requirements.txt
# ===========================================
# DocuMind AI — Dependencies
# ===========================================

# Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20

# LLM & AI
groq==0.15.0
langchain==0.3.14
langchain-groq==0.2.4
langchain-community==0.3.14
langchain-huggingface==0.1.2

# Embeddings (Free, runs locally)
sentence-transformers==3.3.1

# Vector Store
chromadb==0.5.23

# Document Processing
pypdf==5.1.0
python-docx==1.1.2
unstructured==0.16.12
tiktoken==0.8.0

# Database
sqlalchemy==2.0.36
aiosqlite==0.20.0

# Evaluation
ragas==0.2.8
datasets==3.2.0

# Frontend
streamlit==1.41.1

# Utilities
python-dotenv==1.0.1
pydantic==2.10.4
pydantic-settings==2.7.1
loguru==0.7.3
aiofiles==24.1.0
httpx==0.28.1

# Reranking (Free)
rank-bm25==0.2.2
